src/main/java/io/hhplus/tdd/point/PointController.java
package io.hhplus.tdd.point;

import lombok.extern.slf4j.Slf4j;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.web.bind.annotation.*;

import java.util.List;

@Slf4j
@RequestMapping("/point")
@RestController
public class PointController {

    private final PointService pointService;

    @Autowired
    public PointController(PointService pointService) {
        this.pointService = pointService;
    }

    /**
     * 특정 유저의 포인트를 조회
     */
    @GetMapping("{id}")
    public UserPoint point(@PathVariable Long id) {
        log.info("Fetching points for user with ID: {}", id);
        return pointService.getUserPoint(id);
    }

    /**
     * 특정 유저의 포인트 충전/이용 내역을 조회
     */
    @GetMapping("{id}/histories")
    public List<PointHistory> history(@PathVariable Long id) {
        log.info("Fetching point histories for user with ID: {}", id);
        return pointService.getPointHistories(id);
    }

    /**
     * 특정 유저의 포인트를 충전
     */
    @PatchMapping("{id}/charge")
    public UserPoint charge(@PathVariable Long id, @RequestBody Long amount) {
        log.info("Charging points for user with ID: {}, Amount: {}", id, amount);
        return pointService.chargePoint(id, amount);
    }

    /**
     * 특정 유저의 포인트를 사용
     */
    @PatchMapping("{id}/use")
    public UserPoint use(@PathVariable Long id, @RequestBody Long amount) {
        log.info("Using points for user with ID: {}, Amount: {}", id, amount);
        return pointService.usePoint(id, amount);
    }
}
